pyproject.toml
[project]
name = "appsec_crew"
version = "0.1.0"
description = "CrewAI-based AppSec automation for GitHub"
authors = [{ name = "AppSec Crew contributors" }]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.10.1",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
appsec-crew = "appsec_crew.main:main"

[project.urls]
Homepage = "https://github.com/celagus/appsec-crew"
Repository = "https://github.com/celagus/appsec-crew"
Issues = "https://github.com/celagus/appsec-crew/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/appsec_crew"]

[tool.crewai]
type = "crew"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
